# srcs/backend/src/lib/api-zod/src/generated/types/elemental.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * Elemental Duel API — futuristic multiplayer arena game
 * OpenAPI spec version: 0.1.0
 */

export type Elemental = typeof Elemental[keyof typeof Elemental];


export const Elemental = {
  TITAN: 'TITAN',
  RAZOR: 'RAZOR',
  WRAITH: 'WRAITH',
} as const;