# srcs/backend/src/lib/api-zod/src/generated/types/round.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * Elemental Duel API — futuristic multiplayer arena game
 * OpenAPI spec version: 0.1.0
 */
import type { Elemental } from './elemental';
import type { RoundOutcome } from './roundOutcome';

export interface Round {
  id: number;
  matchId: number;
  roundNumber: number;
  player1Choice: Elemental;
  player2Choice: Elemental;
  outcome: RoundOutcome;
  createdAt: Date;
}