# srcs/backend/src/lib/api-zod/src/generated/types/roundInput.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * Elemental Duel API — futuristic multiplayer arena game
 * OpenAPI spec version: 0.1.0
 */
import type { Elemental } from './elemental';

export interface RoundInput {
  elemental: Elemental;
}