# srcs/backend/src/lib/api-zod/src/generated/types/roundOutcome.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * Elemental Duel API — futuristic multiplayer arena game
 * OpenAPI spec version: 0.1.0
 */

export type RoundOutcome = typeof RoundOutcome[keyof typeof RoundOutcome];


export const RoundOutcome = {
  WIN: 'WIN',
  LOSS: 'LOSS',
  DRAW: 'DRAW',
} as const;