# srcs/backend/src/lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * Elemental Duel API — futuristic multiplayer arena game
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  /**
     * @minLength 3
     * @maxLength 20
     */
  username?: string;
  /** @nullable */
  avatarUrl?: string | null;
}